+++ opencti-platform/opencti-front/src/private/components/locations/positions/Root.tsx
@@ -1,42 +1,39 @@
-// TODO Remove this when V6
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck
-import React, { useMemo } from 'react';
-import { Route, Routes, useParams, Link, useLocation, Navigate } from 'react-router-dom';
-import { graphql, usePreloadedQuery, useSubscription } from 'react-relay';
+import React, { useMemo, Suspense } from 'react';
+import { Route, Routes, Link, Navigate, useLocation, useParams } from 'react-router-dom';
+import { graphql, useSubscription, usePreloadedQuery, PreloadedQuery } from 'react-relay';
 import { GraphQLSubscriptionConfig } from 'relay-runtime';
 import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
-import StixCoreObjectContentRoot from '@components/common/stix_core_objects/StixCoreObjectContentRoot';
+import { RootPositionQuery } from '@components/locations/positions/__generated__/RootPositionQuery.graphql';
+import useQueryLoading from 'src/utils/hooks/useQueryLoading';
+import { RootPositionsSubscription } from '@components/locations/positions/__generated__/RootPositionsSubscription.graphql';
 import useForceUpdate from '@components/common/bulk/useForceUpdate';
-import Country from './Country';
-import CountryKnowledge from './CountryKnowledge';
+import StixCoreObjectContentRoot from '../../common/stix_core_objects/StixCoreObjectContentRoot';
+import Position from './Position';
+import PositionKnowledge from './PositionKnowledge';
 import StixDomainObjectHeader from '../../common/stix_domain_objects/StixDomainObjectHeader';
 import FileManager from '../../common/files/FileManager';
+import PositionPopover from './PositionPopover';
+import Loader, { LoaderVariant } from '../../../../components/Loader';
 import StixCoreObjectHistory from '../../common/stix_core_objects/StixCoreObjectHistory';
 import StixCoreObjectOrStixCoreRelationshipContainers from '../../common/containers/StixCoreObjectOrStixCoreRelationshipContainers';
 import StixCoreObjectKnowledgeBar from '../../common/stix_core_objects/StixCoreObjectKnowledgeBar';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
 import EntityStixSightingRelationships from '../../events/stix_sighting_relationships/EntityStixSightingRelationships';
-import { RootCountriesSubscription } from './__generated__/RootCountriesSubscription.graphql';
-import { RootCountryQuery } from './__generated__/RootCountryQuery.graphql';
-import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
-import Loader, { LoaderVariant } from '../../../../components/Loader';
 import { useFormatter } from '../../../../components/i18n';
-import CountryPopover from './CountryPopover';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import { getCurrentTab, getPaddingRight } from '../../../../utils/utils';
-import CountryEdition from './CountryEdition';
+import PositionEdition from './PositionEdition';
 import Security from '../../../../utils/Security';
 import { KNOWLEDGE_KNUPDATE } from '../../../../utils/hooks/useGranted';
 
 const subscription = graphql`
-  subscription RootCountriesSubscription($id: ID!) {
+  subscription RootPositionsSubscription($id: ID!) {
     stixDomainObject(id: $id) {
-      ... on Country {
-        ...Country_country
-        ...CountryEditionOverview_country
+      ... on Position {
+        ...Position_position
+        ...PositionEditionContainer_position
       }
       ...FileImportViewer_entity
       ...FileExportViewer_entity
@@ -46,19 +43,19 @@ const subscription = graphql`
   }
 `;
 
-const countryQuery = graphql`
-  query RootCountryQuery($id: String!) {
-    country(id: $id) {
+const positionQuery = graphql`
+  query RootPositionQuery($id: String!) {
+    position(id: $id) {
       id
+      entity_type
       name
       x_opencti_aliases
-      x_opencti_graph_data
       stixCoreObjectsDistribution(field: "entity_type", operation: count) {
         label
         value
       }
-      ...Country_country
-      ...CountryKnowledge_country
+      ...Position_position
+      ...PositionKnowledge_position
       ...FileImportViewer_entity
       ...FileExportViewer_entity
       ...FileExternalReferencesViewer_entity
@@ -74,27 +71,35 @@ const countryQuery = graphql`
   }
 `;
 
-const RootCountryComponent = ({ queryRef, countryId }) => {
-  const subConfig = useMemo<
-  GraphQLSubscriptionConfig<RootCountriesSubscription>
-  >(
-    () => ({
-      subscription,
-      variables: { id: countryId },
-    }),
-    [countryId],
-  );
-  useSubscription(subConfig);
+type RootPositionProps = {
+  positionId: string;
+  queryRef: PreloadedQuery<RootPositionQuery>;
+};
+
+const RootPosition = ({ positionId, queryRef }: RootPositionProps) => {
+  const subConfig = useMemo<GraphQLSubscriptionConfig<RootPositionsSubscription>>(() => ({
+    subscription,
+    variables: { id: positionId },
+  }), [positionId]);
+
   const location = useLocation();
   const { t_i18n } = useFormatter();
-  const data = usePreloadedQuery(countryQuery, queryRef);
+  useSubscription<RootPositionsSubscription>(subConfig);
+
+  const {
+    position,
+    connectorsForExport,
+    connectorsForImport,
+  } = usePreloadedQuery<RootPositionQuery>(positionQuery, queryRef);
+
   const { forceUpdate } = useForceUpdate();
-  const { country, connectorsForImport, connectorsForExport } = data;
-  const link = `/dashboard/locations/countries/${countryId}/knowledge`;
-  const paddingRight = getPaddingRight(location.pathname, country?.id, '/dashboard/locations/countries');
+
+  const link = `/dashboard/locations/positions/${positionId}/knowledge`;
+  const paddingRight = getPaddingRight(location.pathname, positionId, '/dashboard/locations/positions');
+
   return (
     <>
-      {country ? (
+      {position ? (
         <>
           <Routes>
             <Route
@@ -103,11 +108,11 @@ const RootCountryComponent = ({ queryRef, countryId }) => {
                 <StixCoreObjectKnowledgeBar
                   stixCoreObjectLink={link}
                   availableSections={[
+                    'organizations',
                     'regions',
+                    'countries',
                     'areas',
                     'cities',
-                    'organizations',
-                    'threats',
                     'threat_actors',
                     'intrusion_sets',
                     'campaigns',
@@ -117,7 +122,7 @@ const RootCountryComponent = ({ queryRef, countryId }) => {
                     'tools',
                     'observables',
                   ]}
-                  stixCoreObjectsDistribution={country.stixCoreObjectsDistribution}
+                  stixCoreObjectsDistribution={position.stixCoreObjectsDistribution}
                 />
               }
             />
@@ -125,18 +130,18 @@ const RootCountryComponent = ({ queryRef, countryId }) => {
           <div style={{ paddingRight }}>
             <Breadcrumbs variant="object" elements={[
               { label: t_i18n('Locations') },
-              { label: t_i18n('Countries'), link: '/dashboard/locations/countries' },
-              { label: country.name, current: true },
+              { label: t_i18n('Positions'), link: '/dashboard/locations/positions' },
+              { label: position.name, current: true },
             ]}
             />
             <StixDomainObjectHeader
-              entityType="Country"
+              entityType="Position"
               disableSharing={true}
-              stixDomainObject={country}
-              PopoverComponent={<CountryPopover id={country.id} />}
+              stixDomainObject={position}
+              PopoverComponent={<PositionPopover />}
               EditComponent={(
                 <Security needs={[KNOWLEDGE_KNUPDATE]}>
-                  <CountryEdition countryId={country.id} />
+                  <PositionEdition positionId={position.id} />
                 </Security>
               )}
               enableQuickSubscription={true}
@@ -150,48 +155,48 @@ const RootCountryComponent = ({ queryRef, countryId }) => {
               }}
             >
               <Tabs
-                value={getCurrentTab(location.pathname, country.id, '/dashboard/locations/countries')}
+                value={getCurrentTab(location.pathname, position.id, '/dashboard/locations/positions')}
               >
                 <Tab
                   component={Link}
-                  to={`/dashboard/locations/countries/${country.id}`}
-                  value={`/dashboard/locations/countries/${country.id}`}
+                  to={`/dashboard/locations/positions/${position.id}`}
+                  value={`/dashboard/locations/positions/${position.id}`}
                   label={t_i18n('Overview')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/locations/countries/${country.id}/knowledge/overview`}
-                  value={`/dashboard/locations/countries/${country.id}/knowledge`}
+                  to={`/dashboard/locations/positions/${position.id}/knowledge/overview`}
+                  value={`/dashboard/locations/positions/${position.id}/knowledge`}
                   label={t_i18n('Knowledge')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/locations/countries/${country.id}/content`}
-                  value={`/dashboard/locations/countries/${country.id}/content`}
+                  to={`/dashboard/locations/positions/${position.id}/content`}
+                  value={`/dashboard/locations/positions/${position.id}/content`}
                   label={t_i18n('Content')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/locations/countries/${country.id}/analyses`}
-                  value={`/dashboard/locations/countries/${country.id}/analyses`}
+                  to={`/dashboard/locations/positions/${position.id}/analyses`}
+                  value={`/dashboard/locations/positions/${position.id}/analyses`}
                   label={t_i18n('Analyses')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/locations/countries/${country.id}/sightings`}
-                  value={`/dashboard/locations/countries/${country.id}/sightings`}
+                  to={`/dashboard/locations/positions/${position.id}/sightings`}
+                  value={`/dashboard/locations/positions/${position.id}/sightings`}
                   label={t_i18n('Sightings')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/locations/countries/${country.id}/files`}
-                  value={`/dashboard/locations/countries/${country.id}/files`}
+                  to={`/dashboard/locations/positions/${position.id}/files`}
+                  value={`/dashboard/locations/positions/${position.id}/files`}
                   label={t_i18n('Data')}
                 />
                 <Tab
                   component={Link}
-                  to={`/dashboard/locations/countries/${country.id}/history`}
-                  value={`/dashboard/locations/countries/${country.id}/history`}
+                  to={`/dashboard/locations/positions/${position.id}/history`}
+                  value={`/dashboard/locations/positions/${position.id}/history`}
                   label={t_i18n('History')}
                 />
               </Tabs>
@@ -199,65 +204,65 @@ const RootCountryComponent = ({ queryRef, countryId }) => {
             <Routes>
               <Route
                 path="/"
-                element={<Country countryData={country} />}
+                element={
+                  <Position position={position} />
+                }
               />
               <Route
                 path="/knowledge"
                 element={
-                  <Navigate to={`/dashboard/locations/countries/${countryId}/knowledge/overview`} replace={true} />
-              }
+                  <Navigate to={`/dashboard/locations/positions/${positionId}/knowledge/overview`} replace={true} />
+                }
               />
               <Route
                 path="/knowledge/*"
                 element={
                   <div key={forceUpdate}>
-                    <CountryKnowledge countryData={country} />
+                    <PositionKnowledge position={position} />
                   </div>
-              }
+                }
               />
               <Route
                 path="/content/*"
                 element={
                   <StixCoreObjectContentRoot
-                    stixCoreObject={country}
+                    stixCoreObject={position}
                   />
-              }
+                }
               />
               <Route
                 path="/analyses"
                 element={
-                  <StixCoreObjectOrStixCoreRelationshipContainers
-                    stixDomainObjectOrStixCoreRelationship={country}
-                  />
-              }
+                  <StixCoreObjectOrStixCoreRelationshipContainers stixDomainObjectOrStixCoreRelationship={position} />
+                }
               />
               <Route
                 path="/sightings"
                 element={
                   <EntityStixSightingRelationships
-                    entityId={country.id}
+                    entityId={position.id}
                     entityLink={link}
                     noPadding={true}
                     isTo={true}
                   />
-              }
+                }
               />
               <Route
                 path="/files"
                 element={
                   <FileManager
-                    id={countryId}
+                    id={positionId}
                     connectorsImport={connectorsForImport}
                     connectorsExport={connectorsForExport}
-                    entity={country}
+                    entity={position}
                   />
-              }
+                }
               />
               <Route
                 path="/history"
                 element={
-                  <StixCoreObjectHistory stixCoreObjectId={countryId} />
-              }
+                  <StixCoreObjectHistory stixCoreObjectId={positionId} />
+                }
               />
             </Routes>
           </div>
@@ -268,21 +273,21 @@ const RootCountryComponent = ({ queryRef, countryId }) => {
     </>
   );
 };
-
-const RootCountry = () => {
-  const { countryId } = useParams() as { countryId: string };
-  const queryRef = useQueryLoading<RootCountryQuery>(countryQuery, {
-    id: countryId,
+const Root = () => {
+  const { positionId } = useParams() as { positionId: string; };
+  const queryRef = useQueryLoading<RootPositionQuery>(positionQuery, {
+    id: positionId,
   });
+
   return (
     <>
       {queryRef && (
-        <React.Suspense fallback={<Loader variant={LoaderVariant.container} />}>
-          <RootCountryComponent queryRef={queryRef} countryId={countryId} />
-        </React.Suspense>
+        <Suspense fallback={<Loader variant={LoaderVariant.container} />}>
+          <RootPosition positionId={positionId} queryRef={queryRef} />
+        </Suspense>
       )}
     </>
   );
 };
 
-export default RootCountry;
+export default Root;
